+++ sync-for-expenses/src/codatsyncexpenses/sdkconfiguration.py
@@ -19,9 +19,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = 'prealpha'
-    sdk_version: str = '3.1.0'
-    gen_version: str = '2.146.1'
-    user_agent: str = 'speakeasy-sdk/python 3.1.0 2.146.1 prealpha codat-sync-for-expenses'
+    sdk_version: str = '3.2.0'
+    gen_version: str = '2.155.1'
+    user_agent: str = 'speakeasy-sdk/python 3.2.0 2.155.1 prealpha codat-sync-for-expenses'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: