+++ sync-for-expenses/RELEASES.md
@@ -667,3 +667,13 @@ Based on:
 - [python v3.1.0] sync-for-expenses
 ### Releases
 - [PyPI v3.1.0] https://pypi.org/project/codat-sync-for-expenses/3.1.0 - sync-for-expenses
+
+## 2023-10-16 01:26:25
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.99.2 (2.155.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0] sync-for-expenses
+### Releases
+- [PyPI v3.2.0] https://pypi.org/project/codat-sync-for-expenses/3.2.0 - sync-for-expenses